Fjern deprecated metoder, gi riktigere navn på pakker
