--- src/main/java/no/nav/tag/sykefravarsstatistikk/api/ResourceNotFoundException.java
@@ -1,12 +1,0 @@
-package no.nav.tag.sykefravarsstatistikk.api;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends RuntimeException {
-
-    public ResourceNotFoundException(String msg) {
-        super(msg);
-    }
-}
