+++ src/main/java/no/nav/tag/sykefravarsstatistikk/api/domene/sammenligning/Sykefraværprosent.java
@@ -1,4 +1,4 @@
-package no.nav.tag.sykefravarsstatistikk.api.domene.statistikk;
+package no.nav.tag.sykefravarsstatistikk.api.domene.sammenligning;
 
 import lombok.Value;
