--- src/main/java/no/nav/tag/sykefravarsstatistikk/api/statistikk/Sykefraværprosent2.java
@@ -1,14 +1,0 @@
-package no.nav.tag.sykefravarsstatistikk.api.statistikk;
-
-
-import lombok.Builder;
-import lombok.Data;
-
-import java.math.BigDecimal;
-
-@Deprecated
-@Data
-@Builder
-public class Sykefraværprosent2 {
-    private final BigDecimal land;
-}
