+++ src/test/java/no/nav/tag/sykefravarsstatistikk/api/ApplicationTest.java
@@ -1,6 +1,6 @@
 package no.nav.tag.sykefravarsstatistikk.api;
 
-import no.nav.tag.sykefravarsstatistikk.api.statistikk.OpenSykefraværsstatistikkController;
+import no.nav.tag.sykefravarsstatistikk.api.sammenligning.SammenligningController;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,7 +16,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class ApplicationTest {
 
     @Autowired
-    private OpenSykefraværsstatistikkController controller;
+    private SammenligningController controller;
 
     @Test
     public void contexLoads() {
