+++ src/test/java/no/nav/tag/sykefravarsstatistikk/api/domene/InnloggetSelvbetjeningBrukerTest.java
@@ -2,8 +2,6 @@ package no.nav.tag.sykefravarsstatistikk.api.domene;
 
 import no.nav.tag.sykefravarsstatistikk.api.TestUtils;
 import no.nav.tag.sykefravarsstatistikk.api.altinn.Organisasjon;
-import no.nav.tag.sykefravarsstatistikk.api.domene.InnloggetSelvbetjeningBruker;
-import no.nav.tag.sykefravarsstatistikk.api.domene.Orgnr;
 import no.nav.tag.sykefravarsstatistikk.api.tilgangskontroll.TilgangskontrollException;
 import org.junit.Test;
 
@@ -15,7 +13,7 @@ public class InnloggetSelvbetjeningBrukerTest {
 
     @Test
     public void sjekkTilgang__kaster_ingen_Exception_hvis_bruker_har_tilgang_til_orgnr() {
-        InnloggetSelvbetjeningBruker bruker = getInnloggetSelvbetjeningBruker("12345678901");
+        InnloggetBruker bruker = getInnloggetSelvbetjeningBruker("12345678901");
 
         Organisasjon organisasjon = TestUtils.getOrganisasjon("123456789");
         bruker.setOrganisasjoner(List.of(organisasjon));
@@ -25,14 +23,14 @@ public class InnloggetSelvbetjeningBrukerTest {
 
     @Test(expected = TilgangskontrollException.class)
     public void sjekkTilgang__skal_feile_hvis_bruker_ikke_har_tilgang_til_noen_org() {
-        InnloggetSelvbetjeningBruker bruker = getInnloggetSelvbetjeningBruker("12345678901");
+        InnloggetBruker bruker = getInnloggetSelvbetjeningBruker("12345678901");
 
         bruker.sjekkTilgang(new Orgnr("123456789"));
     }
 
     @Test(expected = TilgangskontrollException.class)
     public void sjekkTilgang__skal_feile_hvis_bruker_ikke_har_tilgang_til_orgnr() {
-        InnloggetSelvbetjeningBruker bruker = getInnloggetSelvbetjeningBruker("12345678901");
+        InnloggetBruker bruker = getInnloggetSelvbetjeningBruker("12345678901");
 
         Organisasjon organisasjon = TestUtils.getOrganisasjon("123456789");
         bruker.setOrganisasjoner(List.of(organisasjon));
